migrate from form to expressive, fix maps merge screen

+++ app/src/main/java/com/aliernfrog/lactool/ui/viewmodel/MapsEditViewModel.kt
@@ -1,9 +1,11 @@
 package com.aliernfrog.lactool.ui.viewmodel
 
 import android.annotation.SuppressLint
+import android.content.ClipData
 import android.content.Context
 import android.util.Log
 import androidx.compose.foundation.ScrollState
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Check
@@ -22,14 +24,17 @@ import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.neverEqualPolicy
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.rememberVectorPainter
-import androidx.compose.ui.platform.LocalClipboardManager
+import androidx.compose.ui.platform.ClipEntry
+import androidx.compose.ui.platform.LocalClipboard
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.unit.Density
+import androidx.compose.ui.unit.dp
 import androidx.lifecycle.ViewModel
 import androidx.navigation.NavController
 import coil.imageLoader
@@ -48,8 +53,10 @@ import com.aliernfrog.lactool.impl.Progress
 import com.aliernfrog.lactool.impl.ProgressState
 import com.aliernfrog.lactool.impl.laclib.MapEditorState
 import com.aliernfrog.lactool.ui.component.ErrorWithIcon
+import com.aliernfrog.lactool.ui.component.VerticalSegmentor
 import com.aliernfrog.lactool.ui.component.createSheetStateWithDensity
-import com.aliernfrog.lactool.ui.component.form.ButtonRow
+import com.aliernfrog.lactool.ui.component.expressive.ExpressiveButtonRow
+import com.aliernfrog.lactool.ui.component.expressive.ExpressiveRowIcon
 import com.aliernfrog.lactool.ui.dialog.DeleteConfirmationDialog
 import com.aliernfrog.lactool.util.Destination
 import com.aliernfrog.lactool.util.extension.removeHtml
@@ -58,6 +65,7 @@ import com.aliernfrog.lactool.util.manager.PreferenceManager
 import com.aliernfrog.toptoast.enum.TopToastColor
 import com.aliernfrog.toptoast.state.TopToastState
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -206,37 +214,55 @@ class MapsEditViewModel(
             },
             options = {
                 val context = LocalContext.current
-                val clipboardManager = LocalClipboardManager.current
+                val clipboard = LocalClipboard.current
+                val scope = rememberCoroutineScope()
+
                 val unused = material.usedBy.isEmpty()
                 var showDeleteDialog by remember { mutableStateOf(false) }
 
-                ButtonRow(
-                    title = stringResource(R.string.mapsMaterials_material_copyUrl),
-                    painter = rememberVectorPainter(Icons.Rounded.ContentCopy)
-                ) {
-                    clipboardManager.setText(AnnotatedString(material.url))
-                    topToastState.showToast(R.string.info_copiedToClipboard, Icons.Rounded.ContentCopy)
-                }
-                ButtonRow(
-                    title = stringResource(R.string.mapsMaterials_material_delete),
-                    description = if (unused) stringResource(R.string.mapsMaterials_unused)
-                    else stringResource(R.string.mapsMaterials_material_delete_description)
-                        .replace("%n", material.usedBy.size.toString()),
-                    painter = rememberVectorPainter(Icons.Rounded.Delete),
-                    contentColor = if (unused) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.error
-                ) {
-                    showDeleteDialog = true
-                }
+                VerticalSegmentor(
+                    {
+                        ExpressiveButtonRow(
+                            title = stringResource(R.string.mapsMaterials_material_copyUrl),
+                            icon = {
+                                ExpressiveRowIcon(
+                                    painter = rememberVectorPainter(Icons.Rounded.ContentCopy)
+                                )
+                            }
+                        ) { scope.launch {
+                            clipboard.setClipEntry(ClipEntry(ClipData.newPlainText(
+                                null, material.url
+                            )))
+                            topToastState.showToast(R.string.info_copiedToClipboard, Icons.Rounded.ContentCopy)
+                        } }
+                    },
+                    {
+                        ExpressiveButtonRow(
+                            title = stringResource(R.string.mapsMaterials_material_delete),
+                            description = if (unused) stringResource(R.string.mapsMaterials_unused)
+                            else stringResource(R.string.mapsMaterials_material_delete_description)
+                                .replace("%n", material.usedBy.size.toString()),
+                            contentColor = if (unused) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.error,
+                            icon = {
+                                ExpressiveRowIcon(
+                                    painter = rememberVectorPainter(Icons.Rounded.Delete),
+                                    containerColor = if (unused) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.error
+                                )
+                            }
+                        ) {
+                            showDeleteDialog = true
+                        }
+                    },
+                    modifier = Modifier.padding(horizontal = 12.dp)
+                )
 
-                if (showDeleteDialog) {
-                    DeleteConfirmationDialog(
-                        name = material.name,
-                        onDismissRequest = { showDeleteDialog = false }
-                    ) {
-                        deleteDownloadableMaterial(material, context)
-                        showDeleteDialog = false
-                        mainViewModel.dismissMediaView()
-                    }
+                if (showDeleteDialog) DeleteConfirmationDialog(
+                    name = material.name,
+                    onDismissRequest = { showDeleteDialog = false }
+                ) {
+                    deleteDownloadableMaterial(material, context)
+                    showDeleteDialog = false
+                    mainViewModel.dismissMediaView()
                 }
             }
         ))
